pyproject.toml
[project]
name = "groundstation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "foxglove-sdk>=0.7.1",
    "foxglove-websocket>=0.1.2",
    "lorarf>=1.4.0",
    "mcap>=1.2.2",
    "mcap-protobuf-support>=0.5.3",
    "opencv-python>=4.11.0.86",
]
